streamlit_app/requirements.txt
streamlit>=1.30.0,<2.0.0
firebase-admin>=6.0.0,<7.0.0
google-cloud-firestore>=2.14.0,<3.0.0
